fix(compute): update vitest.base.config relative path after move
The vitest configs in the moved packages still imported
../../../vitest.base.config from their old depth. After moving into
packages/core/compute/<name>/ they need one more level up.
