+++ packages/core/compute/assistant-e2e/vitest.config.ts
@@ -5,7 +5,7 @@
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import { createConfig } from '../../../vitest.base.config.ts';
+import { createConfig } from '../../../../vitest.base.config.ts';
 
 export default createConfig({
   dirname: typeof __dirname !== 'undefined' ? __dirname : path.dirname(fileURLToPath(import.meta.url)),
